feat: migrate /fire and /flood/stations endpoints to AlertInfoController
- Migrated getResidentsByAddress and getHouseholdsByStations methods from FireStationController to AlertInfoController
- Moved related business logic from FireStationService to AlertInfoService, including the utility method buildFirePersonInfoDTO
- Verified functionality of both endpoints:
  - /fire?address=<address>
  - /flood/stations?stations=<list_of_station_numbers>
  via Postman
