+++ src/main/java/com/openclassrooms/safetynet/safetynetapi/controller/AlertInfoController.java
@@ -1,8 +1,6 @@
 package com.openclassrooms.safetynet.safetynetapi.controller;
 
-import com.openclassrooms.safetynet.safetynetapi.dto.ChildDTO;
-import com.openclassrooms.safetynet.safetynetapi.dto.CoveredPersonsByStationDTO;
-import com.openclassrooms.safetynet.safetynetapi.dto.PersonInfoDto;
+import com.openclassrooms.safetynet.safetynetapi.dto.*;
 import com.openclassrooms.safetynet.safetynetapi.service.AlertInfoService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -126,6 +124,48 @@ public class AlertInfoController {
     @GetMapping("/firestation")
     public ResponseEntity<CoveredPersonsByStationDTO> getPersonsCoveredByStation(@RequestParam("stationNumber") int stationNumber) {
         CoveredPersonsByStationDTO response = alertInfoService.getPersonsCoveredByStation(stationNumber);
+        return ResponseEntity.ok(response);
+    }
+
+    /**
+     * GET endpoint that retrieves residents and medical information for a given address.
+     *
+     * <p>This endpoint returns:
+     * <ul>
+     *   <li>The fire station number serving the specified address</li>
+     *   <li>A list of residents living at that address, each including their name, phone number, age,
+     *       medications, and allergies</li>
+     * </ul>
+     *
+     * <p>Example request: <code>/fire?address=1509 Culver St</code></p>
+     * <p>If no fire station is found for the given address, this endpoint will respond with an appropriate HTTP error status.</p>
+     *
+     * @param address the address to look up
+     * @return a ResponseEntity containing a FireStationResidentsDTO with fire station number
+     * and resident details; returns HTTP 200 OK on success
+     */
+    @GetMapping("/fire")
+    public ResponseEntity<FireStationResidentsDTO> getResidentsByAddress(@RequestParam String address) {
+        FireStationResidentsDTO response = alertInfoService.getResidentsByAddress(address);
+        return ResponseEntity.ok(response);
+    }
+
+    /**
+     * Retrieves a list of households covered by the specified fire stations.
+     *
+     * <p>This endpoint is used in case of flooding to identify all households served by
+     * one or more fire stations. For each address, it returns the list of residents living
+     * there along with their personal and medical information (name, phone, age, medications, allergies).</p>
+     *
+     * <p>Example request: <code>/flood/stations?stations=1,2</code></p>
+     *
+     * @param stations the list of fire station numbers
+     * @return a ResponseEntity containing a list of AddressResidentsDTO,
+     *         each representing an address and its residents with medical details
+     */
+    @GetMapping("/flood/stations")
+    public ResponseEntity<List<AddressResidentsDTO>> getHouseholdsByStations(@RequestParam("stations") List<Integer> stations){
+        List<AddressResidentsDTO> response = alertInfoService.getHouseholdsByStations(stations);
         return ResponseEntity.ok(response);
     }
 }
