+++ src/main/java/com/openclassrooms/safetynet/safetynetapi/service/AlertInfoService.java
@@ -14,9 +14,7 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Log4j2
@@ -270,4 +268,124 @@ public class AlertInfoService {
         return responseDTO;
     }
 
+    private FirePersonInfoDTO buildFirePersonInfoDTO(Person person){
+        MedicalRecord medicalRecord = medicalRecordRepository.getMedicalRecordByFirstNameAndLastName(person.getFirstName(), person.getLastName());
+
+        int age = -1;
+        List<String> medications = Collections.emptyList();
+        List<String> allergies = Collections.emptyList();
+
+        if (medicalRecord != null) {
+            age = AgeUtil.calculateAge(medicalRecord.getBirthdate());
+            medications = medicalRecord.getMedications();
+            allergies = medicalRecord.getAllergies();
+        } else {
+            log.warn("No medical record found for1 {} {}", person.getFirstName(), person.getLastName());
+        }
+
+        FirePersonInfoDTO infoDto = new FirePersonInfoDTO(
+                person.getFirstName(),
+                person.getLastName(),
+                person.getPhone(),
+                age,
+                medications,
+                allergies
+        );
+        return infoDto;
+    }
+
+    /**
+     * Retrieves detailed information about residents living at a specific address,
+     * including their personal details and medical records, as well as the fire station number covering that address.
+     *
+     * <p>This method performs the following operations:
+     * <ul>
+     *   <li>Finds the fire station assigned to the given address.</li>
+     *   <li>Fetches all persons residing at that address.</li>
+     *   <li>Retrieves each person's age, medications, and allergies using their medical records.</li>
+     *   <li>Builds and returns a FireStationResidentsDTO containing the station number and a list of detailed resident info.</li>
+     * </ul>
+     *
+     * @param address the address for which to retrieve residents and fire station information
+     * @return a FireStationResidentsDTO containing the fire station number and a list of residents with their medical details
+     * @throws FireStationNotFoundException if no fire station is found for the provided address
+     */
+    public FireStationResidentsDTO getResidentsByAddress(String address) {
+
+        //1 - Find the fire station using the address
+        FireStation fireStation = fireStationRepository.getFireStationByAddress(address);
+        if (fireStation == null) {
+            log.warn("No fire station found for address {}", address);
+            throw new FireStationNotFoundException("No fire station found for address: " + address);
+        }
+
+        int stationNumber = fireStation.getStation();
+
+        //2 - Find the residents at the address
+        List<Person> residents = personRepository.getPersonByAddress(address);
+        log.info("{} resident(s) found at address {}", residents.size(), address);
+
+        if (residents.isEmpty()) {
+            log.warn("No residents found at address {}", address);
+            // Option 1: Return a DTO with an empty list
+            return new FireStationResidentsDTO(stationNumber, Collections.emptyList());
+            // Option 2:  throw new PersonNotFoundException("No residents found at address: " + address);
+        }
+
+        //3 - Build the detailed list of residents
+        List<FirePersonInfoDTO> detailedResidents = new ArrayList<>();
+        for (Person person : residents) {
+            FirePersonInfoDTO infoDto = buildFirePersonInfoDTO(person);
+            detailedResidents.add(infoDto);
+        }
+/*
+    List<FirePersonInfoDTO> detailedResidents = residents.stream()
+            .map(this::buildFirePersonInfoDTO)
+            .collect(Collectors.toList());
+        */
+        //4 - Return le DTO final
+        return new FireStationResidentsDTO(stationNumber, detailedResidents);
+    }
+
+    /**
+     * Retrieves a list of households (grouped by address) served by the given fire station numbers.
+     *
+     * <p>For each fire station number provided, this method fetches the addresses it covers,
+     * then gathers all residents living at those addresses along with their personal and medical information
+     * (first name, last name, phone, age, medications, and allergies).</p>
+     *
+     * <p>This method is used for the /flood/stations endpoint to support flood-related alerts and preparedness.</p>
+     *
+     * @param stationNumbers the list of fire station numbers to retrieve households for
+     * @return a list of AddressResidentsDTO, each containing an address and the detailed info of its residents
+     */
+    public List<AddressResidentsDTO> getHouseholdsByStations(List<Integer> stationNumbers) {
+        // 1. Retrieve all addresses covered by the requested stations
+        Set<String> addresses = new HashSet<>(); //Set to eliminate duplicated addresses
+        for (Integer stationNumber : stationNumbers) {
+            List<String> addressesForStation = fireStationRepository.getAddressesByStation(stationNumber);
+            addresses.addAll(addressesForStation);
+        }
+
+        // 2. Initialize the result list
+        List<AddressResidentsDTO> result = new ArrayList<>();
+
+        //3. For each address, build the residents with their medical info and add them to the resul
+
+        for (String address : addresses) {
+            List<Person> residents = personRepository.getPersonByAddress(address);
+
+            List<FirePersonInfoDTO> residentDtos = new ArrayList<>();
+            for (Person person : residents) {
+                FirePersonInfoDTO dto = buildFirePersonInfoDTO(person);
+                residentDtos.add(dto);
+            }
+
+            AddressResidentsDTO addressResidentsDTO = new AddressResidentsDTO(address, residentDtos);
+            result.add(addressResidentsDTO);
+        }
+
+        // 4. Return the complete list
+        return result;
+    }
 }
